fix registration and user tests
